Separate out Integrations tab
